feat: finished the navigation. Made profile fragment to start at first for the demonstration of working navigation. Created MainViewModel
