+++ app/src/main/AndroidManifest.xml
@@ -4,13 +4,13 @@
 
     <application
         android:allowBackup="true"
-        android:name=".App"
+        android:name=".app.App"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Demal">
-        <activity android:name=".MainActivity">
+        <activity android:name=".main.view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
